requirements.txt
# Resmi Gazete Takip Botu - Python Bağımlılıkları

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# E-posta Gönderimi
sendgrid>=6.10.0

# Environment Yönetimi
python-dotenv>=1.0.0
